refactor: add utils folder
