+++ node-pubchem/src/src/sync/compound/utils/improveCompound.js
