+++ node-pubchem/src/src/sync/compound/utils/improveCompoundPool.js
